Refactor: mobile
