+++ mobile/README.md
@@ -1,4 +1,4 @@
-# calendar2
+# Calendar
 
 Calendar
