+++ mobile/lib/view/auth/view_signup.dart
@@ -1,8 +1,8 @@
-import 'package:calendar/view/auth/ViewLogin.dart';
-import 'package:calendar/widget/WidgetBody.dart';
-import 'package:calendar/widget/WidgetButton.dart';
-import 'package:calendar/widget/WidgetInput.dart';
-import 'package:calendar/widget/WidgetOAuth.dart';
+import 'package:calendar/view/auth/view_login.dart';
+import 'package:calendar/widget/widget_body.dart';
+import 'package:calendar/widget/widget_button.dart';
+import 'package:calendar/widget/widget_input.dart';
+import 'package:calendar/widget/widget_oauth.dart';
 import 'package:flutter/material.dart';
 
 class ViewSignup extends StatefulWidget {
