+++ mobile/test/widget/widget_input_test.dart
@@ -1,4 +1,4 @@
-import 'package:calendar/widget/WidgetInput.dart';
+import 'package:calendar/widget/widget_input.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
